api/requirements.txt
openai
fastapi
requests
uvicorn
python-multipart
uvicorn[standard]
tortoise-orm
aiomysql